+++ src/test/java/StringAddCalculatorTest.java
@@ -1,4 +1,5 @@
 import org.junit.jupiter.api.Test;
+import stringCalculator.StringAddCalculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
